+++ NEWS.rst
@@ -23,6 +23,8 @@ bundle_adjust (:numref:`bundle_adjust`):
     (:numref:`kaguya_tc_refine_intrinsics`).
   * When optimizing intrinsics, cameras that do not share distortion can
     have different distortion types and sizes. (:numref:`limit_ip`).
+  * Added the option ``--propagate-errors`` to propagate the uncertainties from
+    input cameras to triangulated points (:numref:`ba_error_propagation`).  
   * Added the option ``--weight-image``, to weigh observations based on
     geographic location of triangulated points. (:numref:`limit_ip`). 
   * For ASTER cameras, use the RPC model to find interest points. This does